requirements.txt
flask==2.2.3
requests==2.28.2
Pillow==9.4.0
python-dotenv==1.0.0; python_version < "3.11"